backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.1",
    "pydantic>=2.12.4",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.2.1",
    "google-cloud-aiplatform>=1.126.1",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
]
